+++ app/settings/(accounts)/password/page.tsx
@@ -2,7 +2,7 @@
 
 import { SubmitHandler, useForm } from "react-hook-form";
 import { CURRENT_PASSWORD_RULES, ERROR_MESSAGE, NEW_PASSWORD_RULES, PLACEHOLDER } from "@/constants/inputConstant";
-import * as styles from "@/app/(settingsMenu)/(Account)/password/page.css";
+import * as styles from "@/app/settings/(accounts)/password/page.css";
 import ErrorMessage from "@/components/@common/ErrorMessage";
 
 interface IFormInput {
