+++ app/settings/(accounts)/profile/page.css.ts
