+++ constants/pagesWithHeader.ts
@@ -1,4 +1,4 @@
 export const pagesWithHeader = {
-  logo: ["home/select", "/healthlog/select", "/diary/select", "/faq", "/notice", "/ask", "/settings", "/received-invite"],
-  petGroup: ["/home", "/healthlog", "/healthlog/edit", "/diary/edit", "/diary/detail", "/member", "/invitation"],
+  logo: ["home/select", "/healthlog/select", "/diary/select", "/settings/faq", "/settings/notice", "/settings/ask", "/settings", "/settings/received-invite"],
+  petGroup: ["/home", "/healthlog", "/healthlog/edit", "/diary/edit", "/diary/detail", "/settings/member", "/settings/invitation"],
 };
